[FIX] bug

+++ user/src/main/java/kr/hs/entrydsm/user/integrate/admin/UserExportRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Published
 public interface UserExportRepository {
-    User findByReceiptCode(int receiptCode);
+    User findByReceiptCode(long receiptCode);
     Page<User> findAll(Pageable page, Long receiptCode,
                        boolean isDaejeon, boolean isNationwide,
                        String telephoneNumber, String name,
